Move here small commits from another 'pull request'.
