+++ CarRental.Presentation/Program.cs
@@ -10,10 +10,10 @@ using System.Reflection;
 using CarRental.Presentation.Managers;
 using CarRental.Data.Models.Automobile;
 using CarRental.Data.Models.Checkup;
-using CarRental.Data.Models;
 //using Dapper;   // FOR THE TEST PURPOSES ONLY!
 using CarRental.Data.Models.RecordTypes;
 using CarRental.Data.Dapper;
+using CarRental.Data.Models.Automobile.RecordTypes;
